+++ com.unity.render-pipelines.universal/Runtime/ScriptableRenderer.cs
@@ -205,9 +205,7 @@ namespace UnityEngine.Rendering.Universal
             // Projection flip sign logic is very deep in GfxDevice::SetInvertProjectionMatrix
             // This setup is tailored especially for overlay camera game view
             // For other scenarios this will be overwritten correctly by SetupCameraProperties
-            bool isOffscreen = cameraData.targetTexture != null;
-            bool invertProjectionMatrix = isOffscreen && SystemInfo.graphicsUVStartsAtTop;
-            float projectionFlipSign = invertProjectionMatrix ? -1.0f : 1.0f;
+            float projectionFlipSign = cameraData.IsCameraProjectionMatrixFlipped() ? -1.0f : 1.0f;
             Vector4 projectionParams = new Vector4(projectionFlipSign, near, far, 1.0f * invFar);
             cmd.SetGlobalVector(ShaderPropertyId.projectionParams, projectionParams);